Factor PidInstrumentation out of JfrProfiler
